make the tests happy.. this was an adventure.
error messages showed NPEs on the toTry. The test would fail with an error message that the `eventually` timed out after ~60 retries.

I wasn't able to figure out what was causing the NPE, other than that adding a startSpan would be enough to cause the NPE. Then while looking at the test, I realized that the `when` was trying to match for a specific samRequestContext (namely the empty one that we use for tests). Since the GoogleGroupSyncMonitor code doesn't use an empty samRequestContext, it wasn't matching that.

Partly for fun, partly to keep the same level of validation, I wanted to make sure we kept the validation of the groupToSyncId and policyToSyncId values. This led to a SO answer that needed updating (pending approval) https://stackoverflow.com/questions/38274584/pass-matchers-and-values-using-mockito-using-scala/38275269#38275269

Now the test is happy and we've used a way to validate individual values while using Matchers for the other parameters for mocked functions.