[WOR-637] Add MRG DELETE endpoint and associated tests (#919)
What:

Adds the ability to delete a managed resource group from a billing profile. This is a follow-on to #911 .

Why:

When an Azure billing profile is deleted, we should remove the associated managed resource group record so it can be reused later. If we don't do this, a subsequent billing profile that attempts to use the same MRG will get a 409 and fail out.