[WOR-1518] Add read_job_result action (#1351)
What:

Creates a new workspace action READ_JOB_RESULT that WSM will use to assert a user's ability to inspect a job's results

Why:

Leo cannot query job results in WSM when there is an auth domain present on a workspace. WSM asserts READ which is auth domain constrainable. WSM should check a non-auth domain constrainable action instead.

How:

Creates a READ_JOB_RESULT action on workspaces and assigns it to the owner, writer, reader and application roles (i.e, all roles that are currently able to READ a workspace).
WSM will be made to assert this action in a follow-on PR once this has shipped.