[DCJ-400] #major version bump
At some point in the past, Sam started publishing its clients with semantic versions.
Recently, we enabled Dependabot in jade-data-repo and saw a PR opened with a bad Sam dependency change as a result:
v0.0.241 -> 0.1-9867891-SNAP
The latter is an old prerelease version published as part of a PR commit from before Sam's usage of semver for publishing.
Bumping Sam to a new major version in an effort to move past any old pre-semver publications.